# api/openapi.yaml
openapi: 3.1.0
info:
  title: Rabbit JetStream Management API
  version: 1.0.0
  description: Stable v1 control-plane and operations contract. Resource GET/HEAD requires operator or auditor authentication by default (401 missing/invalid credentials, 403 insufficient role, 404 read_api_disabled when authentication is unconfigured). Authenticated resource operations accept X-RJS-Tenant; it is required when an identity belongs to multiple tenants, defaults to the sole membership otherwise, and is checked before any tenant backend access. Only explicit loopback-bound RJS_LOCAL_DEMO permits anonymous resource reads; session, audit and writes remain authenticated. Probes, metrics, static UI and published contracts remain public and require network-level exposure controls.
servers:
  - url: /
security: [{bearerAuth: []}]
paths:
...
  /api/v1/consumers/refresh:
    post:
      operationId: refreshGlobalConsumers
      description: Operator-only explicit collection. Builds a bounded generation off to the side with at most 10,000 Streams, 100,000 rows, 128 MiB encoded row budget, 16 workers and 30 seconds. Duplicate refreshes coalesce. Any source failure, limit, timeout or declaration revision change prevents publication; a prior generation remains stale.
      responses:
        '200': {description: A complete new generation was published.}
        '202': {description: A refresh is already collecting.}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '503': {description: Collection failed; response reports unavailable or retained stale state without raw backend errors.}
...
  /api/v1/diagnostics/jobs:
    post:
      operationId: createDiagnosticJob
      security: [{bearerAuth: []}]
      description: Creates one explicit instance-local metadata collection for the authenticated operator. No URL, path, command, header or source selection is accepted. Collection is not automatically retried and restart loses every job. One collection may run and at most two jobs are retained per process; each expires ten minutes after creation. Queue/Stream inventories, messages, subscriptions, logs, environment and credentials are excluded.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/DiagnosticJobRequest'}
      responses:
        '202': {description: Collection admitted, headers: {Location: {schema: {type: string}}}, content: {application/json: {schema: {$ref: '#/components/schemas/DiagnosticJob'}}}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '429': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
  /api/v1/diagnostics/jobs/{id}:
    parameters: [{$ref: '#/components/parameters/DiagnosticJobID'}]
    get:
      operationId: getDiagnosticJob
      security: [{bearerAuth: []}]
      description: Creator-only status. Unknown IDs and jobs owned by another identity are indistinguishable. Expired status remains visible until capacity admission reclaims the identity.
      responses: {'200': {description: Current job state, content: {application/json: {schema: {$ref: '#/components/schemas/DiagnosticJob'}}}}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'} }
    delete:
      operationId: cancelDiagnosticJob
      security: [{bearerAuth: []}]
      description: Creator-only cancellation. Ready, partial, failed, cancelled and expired jobs are returned unchanged. Cancellation removes archive access and a late collector cannot republish bytes.
      responses: {'200': {description: Current terminal job state, content: {application/json: {schema: {$ref: '#/components/schemas/DiagnosticJob'}}}}, '401': {$ref: '#/components/responses/Error'}, '403': {$ref: '#/components/responses/Error'}, '404': {$ref: '#/components/responses/Error'}, '503': {$ref: '#/components/responses/Error'} }
  /api/v1/diagnostics/jobs/{id}/download:
    get:
      operationId: downloadDiagnosticJob
      security: [{bearerAuth: []}]
      description: Creator-only download for ready or partial jobs. Range and query requests are rejected. Authorization is audited before bytes are sent. A server write-completion audit does not prove browser receipt or saving.
      parameters: [{$ref: '#/components/parameters/DiagnosticJobID'}]
      responses:
        '200': {description: Bounded metadata ZIP, headers: {Content-Disposition: {schema: {type: string}}, Content-Length: {schema: {type: integer, maximum: 8388608}}}, content: {application/zip: {schema: {type: string, format: binary, maxLength: 8388608}}}}
        '400': {$ref: '#/components/responses/Error'}
        '401': {$ref: '#/components/responses/Error'}
        '403': {$ref: '#/components/responses/Error'}
        '404': {$ref: '#/components/responses/Error'}
        '409': {$ref: '#/components/responses/Error'}
        '503': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer}
  parameters:
...
    DiagnosticJobID: {name: id, in: path, required: true, schema: {type: string, pattern: '^[0-9a-f]{32}$'}}
  responses:
...
    Error: {description: Stable error envelope, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
  schemas:
...
    DiagnosticJobRequest:
      type: object
      additionalProperties: false
      required: [schema, profile]
      properties:
        schema: {type: string, const: rjs.diagnostics-request.v1}
        profile: {type: string, const: metadata-v1}
    DiagnosticManifestEntry:
      type: object
      additionalProperties: false
      required: [source, collectedAt]
      properties:
        source: {type: string}
        file: {type: string}
        collectedAt: {type: string, format: date-time}
        size: {type: integer, minimum: 0, maximum: 4194304}
        sha256: {type: string, pattern: '^[0-9a-f]{64}$'}
        error: {type: string, description: Fixed non-sensitive omission reason; never a raw dependency error.}
    DiagnosticManifest:
      type: object
      additionalProperties: false
      required: [schema, generatedAt, entries]
      properties:
        schema: {type: string, const: rjs.diagnostics-manifest.v1}
        generatedAt: {type: string, format: date-time}
        entries: {type: array, maxItems: 5, items: {$ref: '#/components/schemas/DiagnosticManifestEntry'}}
    DiagnosticJob:
      type: object
      additionalProperties: false
      required: [id, profile, state, createdAt, expiresAt]
      properties:
        id: {type: string, pattern: '^[0-9a-f]{32}$'}
        profile: {type: string, const: metadata-v1}
        state: {type: string, enum: [collecting, ready, partial, failed, cancelled, expired]}
        createdAt: {type: string, format: date-time}
        expiresAt: {type: string, format: date-time}
        finishedAt: {type: string, format: date-time}
        manifest: {$ref: '#/components/schemas/DiagnosticManifest'}
        errorCode: {type: string, enum: [collection_failed, collection_timeout, archive_limit_exceeded, cancelled, expired]}
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: {type: string}
            message: {type: string}
            mutation:
              $ref: '#/components/schemas/MutationAttemptEvidence'
            issues_version:
              type: string
              const: rjs.queue-validation.v1
              description: Present with issues for semantic Queue validation failures only; decoding errors retain text without guessed paths.
            issues:
              type: array
              description: Paths refer to the submitted document before canonical sorting. Group-level constraints can point to a parent object.
              items:
                type: object
                required: [path, code, message]
                properties:
                  path: {type: string, description: JSON Pointer; may identify a missing required field.}
                  code: {type: string, description: Stable semantic rule identifier.}
                  message: {type: string}
    MutationAttemptEvidence:
      type: object
      required: [schemaVersion, scope, phase, resourceEffects]
      description: Optional PUT/DELETE failure evidence for this receiving handler attempt only. Covers Queue/Stream/Consumer and declaration effects; excludes audit/lock metadata and earlier transport replays. Not an idempotency key, rollback assertion or durable original-request outcome. Absence conveys no phase information. Existing status/code/message semantics are unchanged.
      properties:
        schemaVersion: {type: string, const: rjs.mutation-evidence.v1}
        scope: {type: string, const: receiving-attempt}
        phase: {type: string, enum: [audit_intent, backend, audit_outcome]}
        resourceEffects:
          type: string
          enum: [none, possible]
          description: none only for audit_intent failure before backend invocation; possible for backend or audit_outcome failure, including conflicts. Never authorizes automatic retry.
        intentId:
          type: string
          pattern: '^[0-9a-f]{32}$'
          description: Audit intent correlation when generated. Persistence is not guaranteed, including after a lost publish acknowledgment.
      oneOf:
        - properties:
            phase: {const: audit_intent}
            resourceEffects: {const: none}
        - properties:
            phase: {enum: [backend, audit_outcome]}
            resourceEffects: {const: possible}
...